master.cfg: Add architecture targets for kernel builders
This enables the use of the architecture keyword that is now passed as
argument of the underlying scripts. Now builders will use dedicated
configurations file and build directories for the architecture being
tested. Note, that two builders that target the same kernel version
cannot collide. For now, the kernel builders will only be run for the
'amd64' architecture, as it was initially the case.

Signed-off-by: Romain Perier <romain.perier@gmail.com>