master.cfg: define and enable per architecture kernel builders
Currently, kernel are built and tested only for amd64. Other popular
architectures like arm or arm64 are supported by gentoo but not tested
yet. This adds support for the arm architecture for the kernel builders.
For now, a builder is defined for each kernel branch, but also per target.
As we currently support two architectures it means that we will have two
builders for each branch. The name of the builder is suffixed by the
name of the targeted arch. The schedulers will trigger all builders for
the corresponding branch when a change is detected.

Signed-off-by: Romain Perier <romain.perier@gmail.com>